requirements.txt
fastapi==0.110.2
uvicorn[standard]==0.29.0
transformers==4.40.0
aiohttp==3.9.5
pillow==10.3.0
python-multipart==0.0.9
tensorflow==2.16.1
tf-keras==2.16.0
cachetools===5.3.3
